requirements.txt
cycler==0.10.0
decorator==4.3.2
kiwisolver==1.0.1
matplotlib==3.0.3
memory-profiler==0.55.0
networkx==2.2
numpy==1.16.1
Pillow==5.4.1
pkg-resources==0.0.0
psutil==5.5.1
pyparsing==2.3.1
python-dateutil==2.8.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
torch==1.0.1.post2
torchvision==0.2.2.post3
